sensors: bmp: fix a core guideline valiolation
Storing a floating point number (float) in an integer (int32_t)
without a cast is a violation of the C++ core guidelines. Add
a `static_cast<int32_t>()' to resolve the issue.

+++ include/lwiot/bmp280sensor.h
@@ -93,6 +93,6 @@ namespace lwiot
 
 		/* Methods */
 		int32_t computeTemperature(int32_t input) const;
-		float computePressure(int32_t input);
+		int32_t computePressure(int32_t input);
 	};
 }
